refactor(server): extract /api/context-hub/* into a router module
Ninth route-group extraction. 5 handlers (Stage 1 crawl: brains list,
brains/:id, history, analyze, brand/:id) + the context-hub-only
handleQuickStartSynthesis helper (193 lines, the Quick Start brain
synthesis) -> src/server/routes/context-hub.js, mounted
`app.use('/api/context-hub', contextHubRouter)`.

Auth: MIXED -> per-route. brains + analyze use softAuth; brains/:id is
requireAuth; history + brand/:id are open. No mount-level auth.

Deps all already extracted: pool, anthropic + dateContext (llm),
safeParseLLM (llm-json), requireAuth/softAuth/verifyBrandAccess/
SUPER_ADMIN_IDS (auth), forgeScrape/getBrandPageContent/discoverSubpages
(scrape), quickStartTruncate (text) + randomUUID (crypto). The two
domainToName uses are independent inner closures (moved with their
handlers). No shared-helper blocker; QSS has no local-helper cascade.

Test plan:
- node --check server.js + module -> OK
- npm run lint (no-undef) -> clean (caught SUPER_ADMIN_IDS; added)
- module boot-load (import()) -> resolves
- route-inventory guard -> 213 byte-identical (5 paths)
- vitest -> 93/93

Rollback: revert; verbatim multi-span move + helper.
